requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
django-crontab==0.7.1
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
passlib==1.7.4
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.44.0
whitenoise==6.7.0
